Empty offset data handling. Cleanup of streamlit
Moved a bunch of the functions from the streamlit file to the utilities file.

Changed imports at the top because of this too

Fixed issue with offsets not loading in the settings/jsons loading tab (same try except as in "Offsets" tab)
